更新git构建脚本 备份新脚本

+++ setup-devtools/setup_git.sh
@@ -10,7 +10,7 @@ mkdir -p $SETUP_INSTALL_PREFIX/git ;
 cd $SETUP_INSTALL_PREFIX ;
 
 RE2C_VERSION=2.0.3 ;
-GIT_VERSION=2.28.0 ;
+GIT_VERSION=2.29.2 ;
 GIT_LFS_VERSION=2.12.0 ;
 export PATH="$SETUP_INSTALL_PREFIX/bin:$PATH"
 
@@ -23,17 +23,18 @@ fi
 tar -axvf re2c-$RE2C_VERSION.tar.xz ;
 cd re2c-$RE2C_VERSION ;
 ./configure --prefix=$SETUP_INSTALL_PREFIX/re2c/$RE2C_VERSION --with-pic=yes;
-make -j8;
+make -j || make;
 make install;
 
-
-for UPDATE_LNK in $SETUP_INSTALL_PREFIX/re2c/$RE2C_VERSION/bin/*; do
-    UNDATE_LNK_BASENAME="$(basename "$UPDATE_LNK")";
-    if [ -e "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME" ]; then
-        rm -rf "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME";
-    fi
-    ln -sf "$UPDATE_LNK" "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME" ;
-done
+if [[ -e "$SETUP_INSTALL_PREFIX/re2c/$RE2C_VERSION/bin" ]]; then
+    for UPDATE_LNK in $SETUP_INSTALL_PREFIX/re2c/$RE2C_VERSION/bin/*; do
+        UNDATE_LNK_BASENAME="$(basename "$UPDATE_LNK")";
+        if [ -e "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME" ]; then
+            rm -rf "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME";
+        fi
+        ln -sf "$UPDATE_LNK" "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME" ;
+    done
+fi
 
 cd ..;
 
@@ -47,19 +48,20 @@ fi
 tar -axvf git-$GIT_VERSION.tar.xz ;
 cd git-$GIT_VERSION;
 ./configure --prefix=$SETUP_INSTALL_PREFIX/git/$GIT_VERSION --with-curl --with-expat --with-openssl --with-libpcre2 --with-editor=vim ;
-make -j8 all doc;
+make -j all doc || make all doc;
 make install install-doc install-html;
 cd contrib/subtree;
 make install install-doc install-html;
 
-
-for UPDATE_LNK in $SETUP_INSTALL_PREFIX/git/$GIT_VERSION/bin/*; do
-    UNDATE_LNK_BASENAME="$(basename "$UPDATE_LNK")";
-    if [ -e "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME" ]; then
-        rm -rf "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME";
-    fi
-    ln -sf "$UPDATE_LNK" "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME" ;
-done
+if [[ -e "$SETUP_INSTALL_PREFIX/git/$GIT_VERSION/bin" ]]; then
+    for UPDATE_LNK in $SETUP_INSTALL_PREFIX/git/$GIT_VERSION/bin/*; do
+        UNDATE_LNK_BASENAME="$(basename "$UPDATE_LNK")";
+        if [ -e "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME" ]; then
+            rm -rf "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME";
+        fi
+        ln -sf "$UPDATE_LNK" "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME" ;
+    done
+fi
 cd ../../../ ;
 mkdir -p git-lfs;
 cd git-lfs;
@@ -77,12 +79,14 @@ cd git-lfs-v$GIT_LFS_VERSION ;
 tar -axvf ../git-lfs-linux-amd64-v$GIT_LFS_VERSION.tar.gz ;
 env PREFIX=$SETUP_INSTALL_PREFIX/git-lfs/v$GIT_LFS_VERSION ./install.sh ;
 
-for UPDATE_LNK in $SETUP_INSTALL_PREFIX/git-lfs/v$GIT_LFS_VERSION/bin/*; do
-    UNDATE_LNK_BASENAME="$(basename "$UPDATE_LNK")";
-    if [ -e "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME" ]; then
-        rm -rf "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME";
-    fi
-    ln -sf "$UPDATE_LNK" "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME" ;
-done
+if [[ -e "$SETUP_INSTALL_PREFIX/git-lfs/v$GIT_LFS_VERSION/bin" ]]; then
+    for UPDATE_LNK in $SETUP_INSTALL_PREFIX/git-lfs/v$GIT_LFS_VERSION/bin/*; do
+        UNDATE_LNK_BASENAME="$(basename "$UPDATE_LNK")";
+        if [ -e "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME" ]; then
+            rm -rf "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME";
+        fi
+        ln -sf "$UPDATE_LNK" "$SETUP_INSTALL_PREFIX/bin/$UNDATE_LNK_BASENAME" ;
+    done
+fi
 
 cd ../../ ;
